pyproject.toml
[project]
name = "chicago-rstrips"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow~=2.11",
    "boto3",
    "numpy",
    "pandas",
    "psycopg2-binary",
    "requests>=2.32.5",
    "sodapy",
    "python-dotenv",
    "pyarrow",
    "sqlalchemy",
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "geopandas",
    "shapely",
    "matplotlib",
    "folium"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]